+++ test/index.js
@@ -1,12 +1,12 @@
 /* eslint global-require: 0 */
 
 describe('AcyOrt', () => {
-  // require('./plugin')
-  // require('./store')
-  // require('./workflow')
-  // require('./helper')
-  // require('./config')
-  // require('./utility')
+  require('./plugin')
+  require('./store')
+  require('./workflow')
+  require('./helper')
+  require('./config')
+  require('./utility')
   require('./cli')
-  // require('./logger')
+  require('./logger')
 })